feat: updated  swagger docs for tag endpoints
