--- src/WebApi/Config/DocsSetup/Examples/Initiative/TagOdataResponseExample.cs
@@ -1,35 +1,0 @@
-﻿namespace IAVH.BioTablero.CM.WebApi.Config.DocsSetup.Examples.Initiative;
-
-using System.Collections.Generic;
-
-using IAVH.BioTablero.CM.Application.DTOs.Initiatives;
-using IAVH.BioTablero.CM.Application.DTOs.Utils;
-
-using Swashbuckle.AspNetCore.Filters;
-
-using static IAVH.BioTablero.CM.Core.Domain.Utils.Enums.InitiativesEnums;
-
-/// <summary>
-/// Tag OData response example.
-/// </summary>
-public class TagOdataResponseExample : IExamplesProvider<Dictionary<string, object>>
-{
-    /// <summary>
-    /// Get examples for entity.
-    /// </summary>
-    /// <returns>Entity examples.</returns>
-    public Dictionary<string, object> GetExamples() => new()
-    {
-        ["@odata.count"] = 1,
-        ["value"] = new List<TagDto>()
-        {
-            new()
-            {
-                Id = 0,
-                Name = "Tag example",
-                Url = "https://example.com/tag-data",
-                Category = new EnumEntityDto<TagCategory>(TagCategory.PoliticalContext),
-            },
-        },
-    };
-}
