+++ src/WebApi/Config/DocsSetup/Examples/Tag/TagAddRequestExample.cs
@@ -1,4 +1,4 @@
-﻿namespace IAVH.BioTablero.CM.WebApi.Config.DocsSetup.Examples.Initiative;
+﻿namespace IAVH.BioTablero.CM.WebApi.Config.DocsSetup.Examples.Tag;
 
 using IAVH.BioTablero.CM.Application.DTOs.Initiatives;
 using IAVH.BioTablero.CM.Application.DTOs.Utils;
@@ -8,9 +8,9 @@ using Swashbuckle.AspNetCore.Filters;
 using static IAVH.BioTablero.CM.Core.Domain.Utils.Enums.InitiativesEnums;
 
 /// <summary>
-/// Tag response example.
+/// Tag add request example.
 /// </summary>
-public class TagResponseExample : IExamplesProvider<TagDto>
+public class TagAddRequestExample : IExamplesProvider<TagDto>
 {
     /// <summary>
     /// Get examples for entity.
@@ -18,7 +18,6 @@ public class TagResponseExample : IExamplesProvider<TagDto>
     /// <returns>Entity examples.</returns>
     public TagDto GetExamples() => new()
     {
-        Id = 0,
         Name = "Tag example",
         Url = "https://example.com/tag-data",
         Category = new EnumEntityDto<TagCategory>(TagCategory.PoliticalContext),
