+++ src/WebApi/Controllers/Rest/Initiatives/TagCategoryController.cs
@@ -1,14 +1,10 @@
 ﻿namespace IAVH.BioTablero.CM.WebApi.Controllers.Rest.Initiatives;
 
 using IAVH.BioTablero.CM.Application.Interfaces.General;
-using IAVH.BioTablero.CM.WebApi.Config.DocsSetup.Examples.Initiative;
-
+using IAVH.BioTablero.CM.WebApi.Config.DocsSetup.Examples;
 using IAVH.BioTablero.CM.WebApi.Interfaces;
 
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-
-using Swashbuckle.AspNetCore.Filters;
 
 using static IAVH.BioTablero.CM.Core.Domain.Utils.Enums.InitiativesEnums;
 
@@ -20,6 +16,7 @@ using static IAVH.BioTablero.CM.Core.Domain.Utils.Enums.InitiativesEnums;
 [ApiController]
 [Route("[controller]")]
 [Produces("application/json")]
+[ApiConventionType(typeof(CustomApiConventions))]
 public class TagCategoryController(IWebTools webTools,
     IReadEnumeration<TagCategory> entityService) : ControllerBase
 {
@@ -28,10 +25,7 @@ public class TagCategoryController(IWebTools webTools,
     /// </summary>
     /// <returns>Entities list from parameters.</returns>
     [HttpGet]
-    [ProducesResponseType(typeof(TagCategoryResponseExample), StatusCodes.Status200OK)]
-    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-    [SwaggerResponseExample(StatusCodes.Status200OK, typeof(TagCategoryResponseExample))]
-    public IActionResult Get()
+    public IActionResult GetEnumList()
     {
         var response = entityService.GetAll();
         return webTools.CustomResponse(response);
